Refactor App into component
