+++ src/App/App.js
@@ -1,11 +1,11 @@
 import React, { Component } from 'react';
 import GoogleMap from 'google-map-react';
 import autoBind from 'react-autobind';
-import Header from './Header/Header';
-import BreweryMarker from './BreweryMarker/BreweryMarker';
-import BreweryPage from './BreweryPage/BreweryPage';
+import Header from '../Header/Header';
+import BreweryMarker from '../BreweryMarker/BreweryMarker';
+import BreweryPage from '../BreweryPage/BreweryPage';
 import {geolocated} from 'react-geolocated';
-import Breweries from './breweries.json';
+import Breweries from '../breweries.json';
 import ReactGA from 'react-ga';
 import Helmet from 'react-helmet';
